reorganize crypto and device.c to be more based on fido2/
